Change to file
I was having a hard time with making sure that the expenses that I was tracking were being transferred into  the file I created to keep all expenses in one place.  It wasn't recognizing the emojis at first but then I added the 'utf-8' that would help with unicode errors.